requirements.txt
aiohttp>=3.9.0
PyYAML>=6.0.0
websockets>=12.0
